add chart

+++ ListenAndWrite.Web/Models/ScoreViewModel.cs
@@ -9,7 +9,7 @@ namespace ListenAndWrite.Web.Models
     {
         public int Id { set; get; }
         public int AudioId { get; set; }
-        public int AudioScore { get; set; }
+        public float AudioScore { get; set; }
         public DateTime CreateDate { get; set; }
         public virtual AudioViewModel Audio { set; get; }
     }
